+++ js/selectores.js
@@ -1,0 +1,5 @@
+//declarando variables
+export const carrito = document.querySelector('#carrito');// selecionando el div del carrito
+export const vaciarCarritoBtn = document.querySelector('#vaciar-carrito');// donde puedo vaciar todo los productos
+export const listaCursos = document.querySelector('#lista-cursos'); // div donde estan los cursos de los productos
+export const contenedorCarrito = document.querySelector('#lista-carrito tbody');// donde se va almacenar los carritos
